pyproject.toml
[project]
name = "llmops-lab"
version = "0.1.0"
description = "LLMOps learning and experimentation lab"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
]

[project.scripts]
agentreg = "agentops.cli:main"
agentops = "agentops.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "pylint>=2.17.0",
    "black>=23.0.0",
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.pylint.messages_control]
disable = ["C0111", "R0903"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
